requirements.txt
# ReconScout v2.1 — Dependencies
# ──────────────────────────────────────────────────────────────────
# Core functionality works with Python 3.8+ stdlib ONLY.
# Install optional packages for enhanced capabilities.

# ── Optional: Enhanced DNS enumeration ───────────────────────────
dnspython>=2.4.0         # Full DNS: AXFR zone transfer, SRV, SOA records
                         #   pip install dnspython

# ── Optional: Better WHOIS parsing ────────────────────────────────
python-whois>=0.9.0      # Structured WHOIS parsing with field extraction
                         #   pip install python-whois

# ── Optional: YAML config support ─────────────────────────────────
pyyaml>=6.0              # Use config/config.yaml instead of .json
                         #   pip install pyyaml

# ── Optional: SYN scan (requires root privileges) ─────────────────
scapy>=2.5.0             # Low-level packet crafting for SYN/UDP scans
                         #   pip install scapy
                         #   Usage: sudo python reconscout.py target --syn

# ── Install all optional dependencies at once ─────────────────────
#    pip install dnspython python-whois pyyaml scapy
